Rename functions_read to lua_functions
